bug fix in config controller.
The collector task can only be created when the user enters the
ns1 ApiKey.  If the key is saved to the DB, but the task is not created
the user will need to re-enter the key to be able to create the task
as the key is never returned back to the user once saved in the DB.
This commit fixes the promises to ensure that the save process fails
when any step fails. If it is just a transient error, re-submiting
the form should just work.